Move api under pmg.api
