+++ pmg/__init__.py
@@ -142,8 +142,8 @@ import views
 import user_management
 import admin
 
-from pmg.api import api
-app.register_blueprint(api, subdomain='api')
+from pmg.api.v1 import api as api_v1
+app.register_blueprint(api_v1, subdomain='api')
 
-from pmg.api_v2 import api as api_v2
+from pmg.api.v2 import api as api_v2
 app.register_blueprint(api_v2, subdomain='api', url_prefix='/v2')
